\name{reader-internal}

\alias{add.dir.if.not}
\alias{vec.extract.mat}
\alias{assess.dat.type}
\alias{is.ch}
\alias{wc.windows}
\alias{dir.force.slash}
\alias{as.df}

\title{Internal reader Functions}
\description{
  Internal reader functions
}
\details{
  These are not recommended/needed to be called by the user 
}
\keyword{ internal }